Reorganize gadget code
